Data Generation for SIR/Code/requirements.txt
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
google-generativeai==0.5.2
